requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-core
langchain-openai
langchain-pinecone
langchain-huggingface
pinecone
python-dotenv
sentence-transformers
fitz
frontend
unstructured
pdfminer
aiosqlite
langchain-google-genai
pandas
numpy
scikit-learn
openpyxl